move structure classes to separate file
