+++ client/ayon_core/pipeline/create/__init__.py
@@ -18,7 +18,7 @@ from .exceptions import (
     CreatorsRemoveFailed,
     CreatorsOperationFailed,
 )
-
+from .structures import CreatedInstance
 from .utils import (
     get_last_versions_for_instances,
     get_next_versions_for_instances,
@@ -48,10 +48,7 @@ from .creator_plugins import (
     cache_and_get_instances,
 )
 
-from .context import (
-    CreatedInstance,
-    CreateContext
-)
+from .context import CreateContext
 
 from .legacy_create import (
     LegacyCreator,
@@ -64,6 +61,21 @@ __all__ = (
     "DEFAULT_PRODUCT_TEMPLATE",
     "PRE_CREATE_THUMBNAIL_KEY",
     "DEFAULT_VARIANT_VALUE",
+
+    "UnavailableSharedData",
+    "ImmutableKeyError",
+    "HostMissRequiredMethod",
+    "ConvertorsOperationFailed",
+    "ConvertorsFindFailed",
+    "ConvertorsConversionFailed",
+    "CreatorError",
+    "CreatorsCreateFailed",
+    "CreatorsCollectionFailed",
+    "CreatorsSaveFailed",
+    "CreatorsRemoveFailed",
+    "CreatorsOperationFailed",
+
+    "CreatedInstance",
 
     "get_last_versions_for_instances",
     "get_next_versions_for_instances",
@@ -90,7 +102,6 @@ __all__ = (
 
     "cache_and_get_instances",
 
-    "CreatedInstance",
     "CreateContext",
 
     "LegacyCreator",
